requirements.txt
Django==5.1.6
djangorestframework==3.15.2
web3==7.8.0
python-decouple==3.8
django-redis==5.4.0
psycopg2-binary==2.9.10